+++ ao-api/controllers/ecommerce/update_discount_code.go
@@ -107,7 +107,7 @@ func (ec *EcommerceController) UpdateDiscountCode() gin.HandlerFunc {
 			})
 			return
 		}
-		discountCodeId = didRows[0]["id"].(int)
+		discountCodeId = int(didRows[0]["id"].(int64))
 		stripePromotionCodeId := didRows[0]["stripe_promotion_code_id"].(string)
 
 		stripePromotionCode, err := sc.PromotionCodes.Get(stripePromotionCodeId, nil)
